+++ biojava3-structure/src/main/java/org/biojava/bio/structure/jama/LUDecomposition.java
@@ -247,7 +247,7 @@ public class LUDecomposition implements java.io.Serializable {
    public double[] getDoublePivot () {
       double[] vals = new double[m];
       for (int i = 0; i < m; i++) {
-         vals[i] = (double) piv[i];
+         vals[i] = piv[i];
       }
       return vals;
    }
@@ -261,7 +261,7 @@ public class LUDecomposition implements java.io.Serializable {
       if (m != n) {
          throw new IllegalArgumentException("Matrix must be square.");
       }
-      double d = (double) pivsign;
+      double d = pivsign;
       for (int j = 0; j < n; j++) {
          d *= LU[j][j];
       }
